Extract ScriptCompiler to script_compiler.h/.cc
